requirements.txt
# PlacementOps Backend Dependencies
# Core framework
fastapi==0.115.5
uvicorn[standard]==0.32.1

# Database
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# Auth
PyJWT[cryptography]==2.10.1
supabase==2.9.1

# Utilities
python-dotenv==1.0.1
python-multipart==0.0.12
openpyxl==3.1.5
haversine==2.9.0
zipcodes==1.3.0
Jinja2>=3.1.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-mock==3.14.0
httpx==0.27.2
anyio==4.7.0
aiosqlite==0.20.0
